Create AudioCodecBase.h
