+++ src/AudioCodecs/All.h
@@ -9,9 +9,10 @@
 #include "AudioCodecs/CodecFLAC.h"
 #include "AudioCodecs/CodecVorbis.h"
 #include "AudioCodecs/CodecADPCM.h"
-#include "AudioCodecs/CodecCodec2.h"
+//#include "AudioCodecs/CodecCodec2.h"
 #include "AudioCodecs/CodecGSM.h"
 #include "AudioCodecs/CodecMP3Helix.h"
+#include "AudioCodecs/CodecAACHelix.h"
 #include "AudioCodecs/ContainerBinary.h"
 #include "AudioCodecs/CodecADPCMXQ.h"
 #include "AudioCodecs/CodecCopy.h"
@@ -34,7 +35,6 @@
 #include "AudioCodecs/CodecL8.h"
 #include "AudioCodecs/CodecMTS.h"
 #include "AudioCodecs/CodecWavIMA.h"
-#include "AudioCodecs/CodecAACHelix.h"
 #include "AudioCodecs/CodecBasic.h"
 #include "AudioCodecs/CodecG7xx.h"
 #include "AudioCodecs/CodecLC3.h"
