+++ src/AudioCodecs/CodecBasic.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "AudioCodecs/AudioEncoded.h"
+#include "AudioCodecs/AudioCodecsBase.h"
 #include "AudioCodecs/CodecG7xx.h"
 
 namespace audio_tools {
