+++ src/AudioCodecs/CodecCodec2.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "AudioCodecs/AudioEncoded.h"
+#include "AudioCodecs/AudioCodecsBase.h"
 #include "codec2.h"
