+++ src/AudioCodecs/CodecFLAC.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "AudioCodecs/AudioEncoded.h"
+#include "AudioCodecs/AudioCodecsBase.h"
 #include "AudioTools/Buffers.h"
 #include "AudioBasic/Net.h"
 #include "flac.h"
