+++ src/AudioCodecs/CodecL16.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "AudioCodecs/AudioEncoded.h"
+#include "AudioCodecs/AudioCodecsBase.h"
 namespace audio_tools {
 
 /**
