+++ src/AudioCodecs/CodecOpus.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "AudioCodecs/AudioEncoded.h"
+#include "AudioCodecs/AudioCodecsBase.h"
 #include "Print.h"
 #include "opus.h"
