+++ src/AudioCodecs/ContainerBinary.h
@@ -16,7 +16,7 @@
 #pragma once
 #include <string.h>
 
-#include "AudioCodecs/AudioEncoded.h"
+#include "AudioCodecs/AudioCodecsBase.h"
 
 namespace audio_tools {
